Toml.Codec.Combinator.Table

Description

Contains TOML-specific combinators for converting between TOML and user data types.

Tables can be represented in TOML in one of the following ways:

foo =
    { x = ...
    , y = ...
    , ...
    }

Or

[foo]
    x = ...
    y = ...
    ...

Since: 1.3.0.0

Tables

table :: TomlCodec a -> Key -> TomlCodec a

Codec for tables. Use it when when you have nested objects.

Since: 0.2.0

Error Helpers

handleTableErrors :: TomlCodec a -> Key -> TOML -> Validation [TomlDecodeError] a

Maps errors in tables with mapTableErrors

Since: 1.3.0.0

mapTableErrors :: Key -> [TomlDecodeError] -> [TomlDecodeError]

Prepends given key to all errors that contain key. This function is used to give better error messages. So when error happens we know all pieces of table key, not only the last one.

Since: 0.2.0