Toml.Codec.Generic

Description

This module contains implementation of the Generic TOML codec. If your data types are big and nested, and you want to have codecs for them without writing a lot of boilerplate code, you can find this module helpful. Below you can find the detailed explanation on how the Generic codecs work.

Consider the following Haskell data types:

data User = User
    { age     :: Int
    , address :: Address
    , socials :: [Social]
    } deriving (Generic)

data Address = Address
    { street :: Text
    , house  :: Int
    } deriving (Generic)

data Social = Social
    { name :: Text
    , link :: Text
    } deriving (Generic)

Value of the User type represents the following TOML:

age = 27

[address]
  street = "Miami Beach"
  house  = 42

[[socials]]
  name = "Twitter"
  link = "https://twitter.com/foo"

[[socials]]
  name = "GitHub"
  link = "https://github.com/bar"

Normally you would write TomlCodec for this data type like this:

userCodec :: TomlCodec User
userCodec = User
    <$> Toml.int "age" .= age
    <*> Toml.table addressCodec "address" .= address
    <*> Toml.list  socialCodec  "socials" .= socials

addressCodec :: TomlCodec Address
addressCodec = Address
    <$> Toml.text "street" .= street
    <*> Toml.int  "house"  .= house

socialCodec :: TomlCodec Social
socialCodec = Social
    <$> Toml.text "name" .= name
    <*> Toml.text "link" .= link

However, if you derive Generic instance for your data types (as we do in the example), you can write your codecs in a simpler way.

userCodec :: TomlCodec User
userCodec = genericCodec

instance HasCodec Address where
    hasCodec = Toml.table genericCodec

instance HasItemCodec Social where
    hasItemCodec = Right genericCodec

Several notes about the interface:

Your top-level data types are always implemented as genericCodec (or other generic codecs).

If you have a custom data type as a field of another type, you need to implement the instance of the HasCodec typeclass.

If the data type appears as an element of a list, you need to implement the instance of the HasItemCodec typeclass.

Since: 1.1.0.0

Documentation

genericCodec :: (Generic a, GenericCodec (Rep a)) => TomlCodec a

Generic codec for arbitrary data types. Uses field names as keys.

Since: 1.1.0.0

genericCodecWithOptions :: (Generic a, GenericCodec (Rep a), Typeable a) => TomlOptions a -> TomlCodec a

Generic codec with options for arbitrary data types.

Since: 1.1.0.0

stripTypeNameCodec :: (Generic a, GenericCodec (Rep a), Typeable a) => TomlCodec a

Generic codec that uses stripTypeNameOptions.

Since: 1.1.0.0

Options

data TomlOptions (a :: k)

Options to configure various parameters of generic encoding. Specifically:

tomlOptionsFieldModifier: how to translate field names to TOML keys?

Since: 1.1.0.0

Constructors

TomlOptions

Fields

tomlOptionsFieldModifier :: Typeable a => Proxy a -> String -> String

newtype GenericOptions

Same as TomlOptions but with all data type information erased. This data type is used internally. Define your options using TomlOptions data type.

Since: 1.1.0.0

Constructors

GenericOptions

Fields

genericOptionsFieldModifier :: String -> String

stripTypeNameOptions :: forall {k} (a :: k). Typeable a => TomlOptions a

Options that use stripTypeNamePrefix as tomlOptionsFieldModifier.

Since: 1.1.0.0

stripTypeNamePrefix :: forall {k} (a :: k). Typeable a => Proxy a -> String -> String

Strips name of the type name from field name prefix.

>>> data UserData = UserData { userDataId :: Int, userDataShortInfo :: Text }
>>> stripTypeNamePrefix (Proxy @UserData) "userDataId"
"id"
>>> stripTypeNamePrefix (Proxy @UserData) "userDataShortInfo"
"shortInfo"
>>> stripTypeNamePrefix (Proxy @UserData) "udStats"
"stats"
>>> stripTypeNamePrefix (Proxy @UserData) "fooBar"
"bar"
>>> stripTypeNamePrefix (Proxy @UserData) "name"
"name"

Since: 1.1.0.0

Core generic typeclass

class HasCodec a where

Helper typeclass for generic deriving. This instance tells how the data type should be coded if it's a field of another data type.

NOTE: If you implement TOML codecs for your data types manually, prefer more explicit Toml.int or Toml.text instead of implicit Toml.hasCodec. Implement instances of this typeclass only when using genericCodec and when your custom data types are not covered here.

Since: 1.1.0.0

Methods

hasCodec :: Key -> TomlCodec a

class HasItemCodec a where

This typeclass tells how the data type should be coded as an item of a list. Lists in TOML can have two types: primitive and table of arrays.

If hasItemCodec returns Left: primitive arrays codec is used.

If hasItemCodec returns 'Right:' table of arrays codec is used.

Since: 1.1.0.0

Methods

hasItemCodec :: Either (TomlBiMap a AnyValue) (TomlCodec a)

class GenericCodec (f :: k -> Type) where

Helper class to derive TOML codecs generically.

Since: 1.1.0.0

Methods

genericTomlCodec :: forall (p :: k). GenericOptions -> TomlCodec (f p)

ByteString newtypes

There are two ways to encode ByteString in TOML:

Via text.

Via an array of integers (aka array of bytes).

To handle all these cases, tomland provides helpful newtypes, specifically:

ByteStringAsText

ByteStringAsBytes

LByteStringAsText

LByteStringAsBytes

As a bonus, on GHC >= 8.6 you can use these newtypes with the DerivingVia extensions for your own ByteString types.

newtype MyByteString = MyByteString
    { unMyByteString :: ByteString
    } deriving HasCodec via ByteStringAsBytes

newtype ByteStringAsText

Newtype wrapper over ByteString to be used for text values.

Since: 1.3.0.0

Constructors

ByteStringAsText

Fields

unByteStringAsText :: ByteString

newtype ByteStringAsBytes

Newtype wrapper over ByteString to be used for array of integers representation.

Since: 1.3.0.0

Constructors

ByteStringAsBytes

Fields

unByteStringAsBytes :: ByteString

newtype LByteStringAsText

Newtype wrapper over lazy ByteString to be used for text values.

Since: 1.3.0.0

Constructors

LByteStringAsText

Fields

unLByteStringAsText :: ByteString

newtype LByteStringAsBytes

Newtype wrapper over lazy ByteString to be used for array of integers representation.

Since: 1.3.0.0

Constructors

LByteStringAsBytes

Fields

unLByteStringAsBytes :: ByteString

Deriving Via

newtype TomlTable a

newtype for generic deriving of HasCodec typeclass for custom data types that should we wrapped into separate table. Use it only for data types that are fields of another data types.

data Person = Person
    { personName    :: !Text
    , personAddress :: !Address
    } deriving (Generic)

data Address = Address
    { addressStreet :: !Text
    , addressHouse  :: !Int
    } deriving (Generic)
      deriving HasCodec via TomlTable Address

personCodec :: TomlCodec Person
personCodec = stripTypeNameCodec

personCodec corresponds to the TOML of the following structure:

name = "foo"
[address]
    addressStreet = "Bar"
    addressHouse = 42

Since: 1.3.0.0

Constructors

TomlTable

Fields

unTomlTable :: a

newtype TomlTableStrip a

newtype for generic deriving of HasCodec typeclass for custom data types that should be wrapped into a separate table.

Similar to TomlTable but also strips the data type name prefix from TOML keys.

personCodec from the TomlTable comment corresponds to the TOML of the following structure:

name = "foo"
[address]
    street = "Bar"
    house = 42

Since: 1.3.2.0

Constructors

TomlTableStrip

Fields

unTomlTableStrip :: a