Toml.Codec.Types

Description

Contains general underlying monad for bidirectional conversion.

Since: 1.3.0.0

Toml Codec

type TomlCodec a = Codec a a

Specialied Codec type alias for bidirectional TOML serialization. Keeps TOML object as both environment and state.

Since: 0.5.0

Toml Environment

type TomlEnv a = TOML -> Validation [TomlDecodeError] a

Immutable environment for TOML conversion.

Since: 1.3.0.0

Toml State

newtype TomlState a

Mutable context for TOML conversion. We are introducing our own implemetation of state with MonadState instance due to some limitation in the design connected to the usage of State.

This newtype is equivalent to the following transformer:

MaybeT (State TOML)

Since: 1.3.0.0

Constructors

TomlState

Fields

unTomlState :: TOML -> (Maybe a, TOML)

eitherToTomlState :: Either e a -> TomlState a

Transform Either into TomlState.

Since: 1.3.0.0

Codec

data Codec i o

Monad for bidirectional conversion. Contains pair of functions:

How to read value of type o (out) from immutable environment context (TomlEnv)?

How to store a value of type i (in) in stateful context (TomlState) and return a value of type o?

This approach with the bunch of utility functions allows to have single description for from/to TOML conversion.

In practice this type will always be used in the following way:

type TomlCodec a = Codec a a

Type parameter i if fictional. Here some trick is used. This trick is implemented in the codec package and described in more details in related blog post: https://blog.poisson.chat/posts/2016-10-12-bidirectional-serialization.html.

Since: 0.0.0

Constructors

Codec

Fields

codecRead :: TomlEnv o

Extract value of type o from monadic context TomlEnv.

codecWrite :: i -> TomlState o

Store value of type i inside monadic context TomlState and returning value of type o. Type of this function actually should be o -> TomlState () but with such type it's impossible to have Monad and other instances.

Function alternative

(<!>) :: Alternative f => (a -> f x) -> (a -> f x) -> a -> f x infixl 3

Alternative instance for function arrow but without empty.