Toml.Parser.Value

Description

Parser for UValue.

Documentation

arrayP :: Parser [UValue]

Parser for array of values. This parser tries to parse first element of array, pattern-matches on this element and uses parser according to this first element. This allows to prevent parsing of heterogeneous arrays.

boolP :: Parser Bool

Parser for Bool value.

dateTimeP :: Parser UValue

Parser for datetime values.

doubleP :: Parser Double

Parser for Double value.

integerP :: Parser Integer

Parser for Integer value.

valueP :: Parser UValue

Parser for UValue.

anyValueP :: Parser AnyValue

Uses valueP and typechecks it.