Toml.Type.Key

Description

Implementation of key type. The type is used for key-value pairs and table names.

Since: 1.3.0.0

Core types

newtype Key

Key of value in key = val pair. Represents as non-empty list of key components — Pieces. Key like

site."google.com"

is represented like

Key (Piece "site" :| [Piece "\"google.com\""])

Since: 0.0.0

Constructors

Key

Fields

unKey :: NonEmpty Piece

type Prefix = Key

Type synonym for Key.

Since: 0.0.0

newtype Piece

Represents the key piece of some layer.

Since: 0.0.0

Constructors

Piece

Fields

unPiece :: Text

pattern (:||) :: Piece -> [Piece] -> Key

Bidirectional pattern synonym for constructing and deconstructing Keys.

(<|) :: Piece -> Key -> Key

Prepends Piece to the beginning of the Key.

Key difference

data KeysDiff

Data represent difference between two keys.

Since: 0.0.0

Constructors

Equal
Keys are equal

NoPrefix
Keys don't have any common part.

FstIsPref
The first key is the prefix of the second one.

Fields

!Key

Rest of the second key.

SndIsPref
The second key is the prefix of the first one.

Fields

!Key

Rest of the first key.

Diff
Key have a common prefix.

Fields

!Key

Common prefix.

!Key

Rest of the first key.

!Key

Rest of the second key.

keysDiff :: Key -> Key -> KeysDiff

Find key difference between two keys.

Since: 0.0.0